railwaymcq/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
sqlparse==0.4.4
tzdata==2024.1
